feat: Create MainPage
- MainPage 컴포넌트 구현
- Chat 컴포넌트에서 프롭스 의존성 제거
- Main 컴포넌트 제거
