+++ src/components/Chat/Chat/Chat.module.css
@@ -1,25 +1,12 @@
-.chatHeader {
-  display: flex;
-  align-items: center;
-  justify-content: flex-start; /* 왼쪽 정렬 */
-  width: 100%;
-  min-height: 56px;
-  position: sticky;
-  padding-left: 20px;
-  gap: 8px;
-}
-
 .chatContainer {
   position: relative;
   display: flex;
   flex-direction: column;
   justify-content: flex-start; /* 위쪽부터 시작하도록 설정 */
   align-items: center;
-  position: relative;
-  height: 100vh; /* 전체 높이를 100vh로 설정 */
-  margin: 0 auto; /* 가운데 정렬 */
+  margin: 0 auto;
   flex: 1;
-  transition: margin-left 0.3s ease;
+  height: 100%;
 }
 
 .chatMessages {
@@ -27,6 +14,7 @@
   flex-direction: column;
   align-items: center;
   width: 100%;
+  overflow-x: hidden;
   overflow-y: auto;
   padding-bottom: 70px; /* QuestionBox 영역과 겹치지 않도록 여백 추가 */
   box-sizing: border-box; /* 패딩을 포함하여 크기를 계산 */
@@ -41,94 +29,19 @@
 }
 
 /* 커스텀 스크롤바 스타일링 */
-.chat-messages::-webkit-scrollbar {
-  width: 1rem; /* 스크롤바의 너비 */
+.chatMessages::-webkit-scrollbar {
+  width: 8px; /* 스크롤바의 너비 */
 }
 
-.chat-messages::-webkit-scrollbar-thumb {
+.chatMessages::-webkit-scrollbar-thumb {
   background-color: rgba(122, 122, 122, 0.36); /* 스크롤바의 색상 */
   border-radius: 1rem; /* 스크롤바의 라운드 */
 }
 
-.chat-messages::-webkit-scrollbar-track {
+.chatMessages::-webkit-scrollbar-track {
   background: transparent; /* 트랙 배경을 투명하게 */
 }
 
-.chat-messages::-webkit-scrollbar-button {
+.chatMessages::-webkit-scrollbar-button {
   display: none; /* 스크롤바 화살표 제거 */
-}
-
-.sidebarVisible {
-  margin-left: 300px;
-}
-
-.sidebarHidden {
-  margin-left: 0;
-}
-
-.plusIcon {
-  position: relative;
-  width: 14px;
-  height: 14px;
-}
-
-.plusIcon::before,
-.plusIcon::after {
-  content: '';
-  position: absolute;
-  background-color: #b1b1b1;
-}
-
-.plusIcon::before {
-  width: 2px;
-  height: 100%;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-}
-
-.plusIcon::after {
-  width: 100%;
-  height: 2px;
-  top: 50%;
-  left: 0;
-  transform: translateY(-50%);
-}
-
-.IconButton {
-  background: transparent;
-  border: none;
-  cursor: pointer;
-}
-
-.sidebarVisible .IconButton {
-  display: none;
-}
-.sidebarHidden .IconButton {
-  display: block;
-}
-
-@media only screen and (max-width: 768px) {
-  /*모바일*/
-  .chatHeader {
-    width: 85%;
-    justify-content: space-between;
-    padding-left: 0;
-    gap: 0;
-  }
-
-  .chatHeader > :nth-child(2) {
-    order: 3;
-  }
-  .chatHeader > :nth-child(3) {
-    order: 2;
-  }
-
-  .sidebarVisible {
-    margin-left: 0;
-  }
-
-  .sidebarVisible .IconButton {
-    display: block;
-  }
 }
