--- src/pages/Main/Main.module.css
@@ -1,16 +1,0 @@
-.container {
-  width: 100%;
-  height: 100vh;
-}
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background: rgba(0, 0, 0, 0.5);
-  opacity: 1;
-  visibility: visible;
-  transition: opacity 0.3s ease;
-  z-index: 1;
-}
