+++ src/pages/MainPage/MainPage.module.css
@@ -1,0 +1,38 @@
+.container {
+  width: 100%;
+  height: 100vh;
+  overflow-x: hidden;
+  overflow-y: hidden;
+}
+.overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  opacity: 1;
+  visibility: visible;
+  transition: opacity 0.3s ease;
+  z-index: 99;
+}
+.childrenContainer {
+  height: calc(100vh - 56px); 
+  margin-top: 56px;
+  transition: margin-left 0.3s ease;
+}
+
+.sidebarVisible {
+  margin-left: 300px;
+}
+
+.sidebarHidden {
+  margin-left: 0;
+}
+
+@media only screen and (max-width: 768px) {
+  /*모바일*/
+  .sidebarVisible {
+    margin-left: 0;
+  }
+}
